fix u_noise in connectivity, fix init_weights in network and run_onelayer

+++ dev_ori_sel_RF/run_onelayer.py
@@ -114,8 +114,8 @@ def parameter_sweep_layer4(Version,config_dict,**kwargs):
 
                     "init_weights" : tf.convert_to_tensor(init_weights,dtype=tf.float32)
                         if config_dict["Wlgn_to4_params"]["mult_norm"]!="xalpha_approx"
-                        else [tf.convert_to_tensor(init_weights[0],dtype=tf.float64),#float32
-                            tf.convert_to_tensor(init_weights[1],dtype=tf.float64)], #float32
+                        else [tf.convert_to_tensor(init_weights[0],dtype=tf.float32),#float32
+                            tf.convert_to_tensor(init_weights[1],dtype=tf.float32)], #float32
                     "Wret_to_lgn" : tf.convert_to_tensor(Wret_to_lgn,dtype=tf.float32),
                     "W4to4" : tf.convert_to_tensor(W4to4, dtype=tf.float32),
                     "W23to23" : tf.convert_to_tensor(np.array([]), dtype=tf.float32),
